pebblesdk_sys
Function action_bar_layer_add_to_window

pub unsafe extern "C" fn action_bar_layer_add_to_window(
    action_bar: *mut ActionBarLayer,
    window: *mut Window,
)

Adds the action bar’s layer on top of the window’s root layer. It also adjusts the layout of the action bar to match the geometry of the window it gets added to. Lastly, it calls window_set_click_config_provider_with_context on the window to set it up to work with the internal callback and raw click handlers of the action bar, to enable the highlighting of the section of the action bar when the user presses a button.

After this call, do not use window_set_click_config_provider_with_context with the window that the action bar has been added to (this would de-associate the action bar’s click config provider and context). Instead use action_bar_layer_set_click_config_provider and action_bar_layer_set_context to register the click configuration provider to configure the buttons actions.

It is advised to call this is in the window’s .load or .appear handler. Make sure to call action_bar_layer_remove_from_window in the window’s .unload or .disappear handler.

Adding additional layers to the window’s root layer after this calll can occlude the action bar.