pebblesdk_sys
Function action_bar_layer_set_icon_animated

pub unsafe extern "C" fn action_bar_layer_set_icon_animated(
    action_bar: *mut ActionBarLayer,
    button_id: ButtonId,
    icon: *const GBitmap,
    animated: bool,
)

Sets an action bar icon onto one of the 3 slots as identified by button_id. Only BUTTON_ID_UP, BUTTON_ID_SELECT and BUTTON_ID_DOWN can be used. Optionally, if animated is true, the transition will be animated. Whenever an icon is set, the click configuration provider will be called, to give the application the opportunity to reconfigure the button interaction.

See also action_bar_layer_set_icon(), action_bar_layer_set_icon_press_animation(), action_bar_layer_set_click_config_provider().