pebblesdk_sys
Function app_message_open

pub unsafe extern "C" fn app_message_open(
    size_inbound: u32,
    size_outbound: u32,
) -> AppMessageResult

Open AppMessage to transfers.

Use dict_calc_buffer_size_from_tuplets or dict_calc_buffer_size to estimate the size you need.

It is recommended that if the Inbox will be used, that at least the Inbox callbacks should be registered before this call. Otherwise it is possible for an Inbox message to be NACK’ed without being seen by the application. Returns a result code such as APP_MSG_OK or APP_MSG_OUT_OF_MEMORY.