pebblesdk_sys
Function gbitmap_create_as_sub_bitmap

pub unsafe extern "C" fn gbitmap_create_as_sub_bitmap(
    base_bitmap: *const GBitmap,
    sub_rect: GRect,
) -> *mut GBitmap

Create a new GBitmap on the heap as a sub-bitmap of a ‘base’ [``]GBitmap, using a GRect to indicate what portion of the base to use. The sub-bitmap will just reference the image data and palette of the base bitmap. No deep-copying occurs as a result of calling this function, thus the caller is responsible for making sure the base bitmap and palette will remain available when using the sub-bitmap. Note that you should not destroy the parent bitmap until the sub_bitmap has been destroyed. The resulting GBitmap must be destroyed using gbitmap_destroy. Returns a pointer to the GBitmap. NULL if the GBitmap could not be created.