pebblesdk_sys
Function gbitmap_create_with_data

pub unsafe extern "C" fn gbitmap_create_with_data(
    data: *const u8,
) -> *mut GBitmap

Creates a new GBitmap on the heap initialized with the provided Pebble image data.

The resulting GBitmap must be destroyed using gbitmap_destroy but the image data will not be freed automatically. The developer is responsible for keeping the image data in memory as long as the bitmap is used and releasing it after the bitmap is destroyed.

One way to generate Pebble image data is to use bitmapgen.py in the Pebble SDK to generate a .pbi file. Returns a pointer to the GBitmap. NULL if the GBitmap could not be created.