pebblesdk_sys
Function gbitmap_get_bytes_per_row

pub unsafe extern "C" fn gbitmap_get_bytes_per_row(
    bitmap: *const GBitmap,
) -> u16

Get the number of bytes per row in the bitmap data for the given GBitmap. On rectangular displays, this can be used as a safe way of iterating over the rows in the bitmap, since bytes per row should be set according to format. On circular displays with pixel format of GBitmapFormat8BitCircular this will return 0, and should not be used for iteration over frame buffer pixels. Instead, use GBitmapDataRowInfo, which provides safe minimum and maximum x values for a given row’s y value. Returns the number of bytes per row of the GBitmap.

See also gbitmap_get_data.