pebblesdk_sys
Function graphics_draw_bitmap_in_rect

pub unsafe extern "C" fn graphics_draw_bitmap_in_rect(
    ctx: *mut GContext,
    bitmap: *const GBitmap,
    rect: GRect,
)

Draws a bitmap into the graphics context, inside the specified rectangle

If the size of rect is smaller than the size of the bitmap, the bitmap will be clipped on right and bottom edges. If the size of rect is larger than the size of the bitmap, the bitmap will be tiled automatically in both horizontal and vertical directions, effectively drawing a repeating pattern.

See also GBitmap, GContext.