pebblesdk_sys
Function property_animation_update_int16

pub unsafe extern "C" fn property_animation_update_int16(
    property_animation: *mut PropertyAnimation,
    distance_normalized: u32,
)

Default update callback for a property animations to update a property of type int16_t. Assign this function to the .base.update callback field of your PropertyAnimationImplementation, in combination with a .getter and .setter accessors of types Int16Getter and Int16Setter. The implementation of this function will calculate the next value of the animation and call the setter to set the new value upon the subject.

This function is not supposed to be called “manually”, but will be called automatically when the animation is being run.