pebblesdk_sys
Function status_bar_layer_create

pub unsafe extern "C" fn status_bar_layer_create() -> *mut StatusBarLayer

Creates a new StatusBarLayer on the heap and initializes it with the default values.

Text color: [GColorBlack]

Background color: [GColorWhite]

Frame: GRect(0, 0, screen_width, STATUS_BAR_LAYER_HEIGHT) The status bar is automatically marked dirty after this operation. You can call layer_set_frame to create a StatusBarLayer of a different width.

// Change the status bar width to make space for the action bar
int16_t width = layer_get_bounds(root_layer).size.w - ACTION_BAR_WIDTH;
GRect frame = GRect(0, 0, width, STATUS_BAR_LAYER_HEIGHT);
layer_set_frame(status_bar_layer_get_layer(status_bar), frame);
layer_add_child(root_layer, status_bar_layer_get_layer(status_bar));
``` Returns a pointer to the StatusBarLayer, which will be allocated to the heap,
`NULL` if the StatusBarLayer could not be created.